Issue: Resolve conflict between main and        redo-splitting-components branch Solution: Create a separate branch to resolve conflicts Risk: Everyone should pull from main again once this branch is merged Reviewed by: Everyone
