--- frontend/src/components/landingPage/landingPage.module.css
@@ -1,73 +1,0 @@
-.whole_wrapper {
-  width: 100%;
-  height: 100%;
-  background-color: #9ab99d;
-  display: flex;
-  flex-direction: column;
-}
-
-.header_wrapper {
-  height: 200px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.title {
-  font-size: 80px;
-  font-weight: 500;
-}
-
-.body_wrapper {
-  height: 300px;
-  padding: 30px;
-}
-
-.body_content {
-  background-color: #c9dbcc;
-  padding: 30px;
-  font-size: 30px;
-  height: 100%;
-  width: 100%;
-  border-radius: 15px;
-}
-
-.bottom_wrapper {
-  position: relative;
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.play_btn {
-  border: none;
-  background-color: #cf8847;
-  font-size: 50px;
-  font-weight: bold;
-  width: 350px;
-  height: 70px;
-  transition: all 0.3s;
-  border-radius: 15px;
-}
-
-.play_btn:hover {
-  background-color: #bc6a21;
-}
-
-.play_btn:focus {
-  outline: none;
-}
-
-.logo_wrapper {
-  position: absolute;
-  right: 30px;
-}
-
-.logo {
-  width: 80px;
-  margin: 5px;
-  padding: 5px;
-  background-color: white;
-  border-radius: 50%;
-}
