+++ frontend/src/components/logo/logo.module.css
@@ -1,0 +1,7 @@
+.logo {
+  width: 80px;
+  margin: 5px;
+  padding: 5px;
+  background-color: white;
+  border-radius: 50%;
+}
