+++ frontend/src/pages/gamePage/gamePage.jsx
@@ -1,9 +1,8 @@
 import React, { useState, useEffect } from "react";
 import styles from "./gamePage.module.css";
-import Timer from "../timer/timer";
-import TimerBar from "../timerBar/timerBar";
-import TimeBar from "../timeBar/timeBar";
-import TimerGroup from "../timerGroup/timerGroup";
+import Timer from "../../components/timer/timer";
+import TimerBar from "../../components/timeBar/timeBar";
+import TimerGroup from "../../components/timerGroup/timerGroup";
 
 // eslint-disable-next-line
 const GamePage = (props) => {
@@ -20,7 +19,6 @@ const GamePage = (props) => {
           <div className={styles.body_wrapper}>
             <TimerBar sec={sec} />
             <Timer />
-            <TimeBar />
             <TimerGroup />
           </div>
         </p>
