+++ frontend/src/pages/landingPage/landingPage.jsx
@@ -1,7 +1,8 @@
 import React from "react";
 import styles from "./landingPage.module.css";
-import uoaLogo from "../../assets/images/uoa-logo.png";
-import wdccLogo from "../../assets/images/wdcc-logo.png";
+import Buttons, { PlayButton } from "../../components/buttons/button";
+import Logos from "../../components/logo/logo";
+import Textbox, { LandingTextbox } from "../../components/textbox/textbox";
 
 /**
  * This component is the Landing Page for our app
@@ -16,25 +17,18 @@ const LandingPage = ({ switchToConsentPage }) => (
       <h1 className={styles.title}>Spot the Stick Insect!</h1>
     </header>
     <div className={styles.body_wrapper}>
-      <p className={styles.body_content}>
-        In this game, pretend you are a predator hunting for stick insects!
+      <LandingTextbox
+        text="In this game, pretend you are a predator hunting for stick insects!
         Stick insects are masters of camouflage, who use their appearance to
         avoid being caught. You need to find them as quickly as possible by
-        clicking on them!
-      </p>
+        clicking on them!"
+      />
     </div>
 
     <div className={styles.bottom_wrapper}>
-      <button
-        onClick={switchToConsentPage}
-        className={styles.play_btn}
-        type="button"
-      >
-        Play Now
-      </button>
+      <PlayButton func={switchToConsentPage} text="Play Now" />
       <div className={styles.logo_wrapper}>
-        <img src={uoaLogo} alt="uoa logo" className={styles.logo} />
-        <img src={wdccLogo} alt="wdcc logo" className={styles.logo} />
+        <Logos />
       </div>
     </div>
   </section>
