+++ frontend/src/pages/landingPage/landingPage.module.css
@@ -1,0 +1,37 @@
+.whole_wrapper {
+  width: 100%;
+  height: 100%;
+  background-color: #9ab99d;
+  display: flex;
+  flex-direction: column;
+}
+
+.header_wrapper {
+  height: 200px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.title {
+  font-size: 80px;
+  font-weight: 500;
+}
+
+.body_wrapper {
+  height: 300px;
+  padding: 30px;
+}
+
+.bottom_wrapper {
+  position: relative;
+  flex: 1;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.logo_wrapper {
+  position: absolute;
+  right: 30px;
+}
